requirements.txt
fastapi
uvicorn
pinecone
google-genai
python-dotenv
discord.py
python-dotenv
protobuf
pinecone-client
sentence_transformers
numpy